Refactor TrieKey out of utils
