+++ core/primitives/Cargo.toml
@@ -25,6 +25,7 @@ log = "0.4"
 reed-solomon-erasure = "4"
 jemallocator = { version = "0.3.0", optional = true }
 actix = "0.9.0"
+hex = "0.4"
 
 borsh = "0.6.1"
