+++ core/primitives/src/lib.rs
@@ -22,6 +22,7 @@ pub mod state_record;
 pub mod telemetry;
 pub mod test_utils;
 pub mod transaction;
+pub mod trie_key;
 pub mod types;
 pub mod utils;
 pub mod validator_signer;
