+++ test-utils/state-viewer/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 
 [dependencies]
 clap = "2.32.0"
-hex = "0.4"
 ansi_term = "0.12.0"
 borsh = "0.6.1"
